Changed a lot of ints to size_t in for loops. Also made some index-\based loops range-based instead.

+++ src/Model.cpp
@@ -120,10 +120,10 @@ void Model::loadMesh(aiMesh* mesh, const aiScene* scene)
 
     // Get rid of this if we ever make collision detection
     // We're calculating the greatest distance between vectors in the model
-    for (int i {0}; i < vertices.size(); i+=8)
+    for (size_t i {0}; i < vertices.size(); i+=8)
     {
         const glm::vec3 vertex1 {vertices[i], vertices[i+1], vertices[i+2]};
-        for (int j {0}; i < vertices.size(); ++i)
+        for (size_t j {0}; i < vertices.size(); ++i)
         {
             const glm::vec3 vertex2 {vertices[j], vertices[j+1], vertices[j+2]};
             const GLfloat displacementVectorLength {glm::length(vertex1 - vertex2)};
